requirements.txt
Flask==2.3.3
google-genai>=0.4.0
gunicorn==21.2.0
Flask-HTTPAuth==4.8.0
python-dotenv==1.0.0
psutil==5.9.5
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.0
werkzeug==2.3.7
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
requests>=2.31.0
python-docx>=0.8.11
python-pptx>=0.6.21
vertexai>=1.38.0 